fix(config): streamable_http fails closed at load-time on dev DSN
Before this change, `config.Load()` accepted
`MCP_TRANSPORT=streamable_http` paired with any dev
control-plane DSN (empty / "memory" / "file://..." / bare path)
and silently booted against an in-process store. Session state,
audit events, and rate-limit counters diverge across replicas
on that path — the hazard is a multi-process production
deployment that appears to work but loses data invisibly.

`internal/runtime.BuildStore` already rejected this combo, but
only after Load() succeeded; the error surfaced at server-start
rather than at config-load. The transport/auth matrix test
explicitly documented the gap with a "Wave C" TODO comment.

This commit closes it:

- New `internal/config/IsDevControlPlaneDSN` consolidates the
  predicate (memory / empty / file:// / bare path). Replaces
  the runtime-side copy so there is a single source of truth.
- `config.Load()` fails closed when
  `MCP_TRANSPORT=streamable_http` + dev DSN +
  `MCP_ALLOW_DEV_BACKEND != "1"`. Error message names both
  escape hatches (set the flag or use postgres://).
- `runtime.BuildStore` now calls `config.IsDevControlPlaneDSN`
  as defence-in-depth; any caller that bypasses Load() still
  refuses to start against a dev DSN.
- The unreachable `cfg.ControlPlaneDSN == ""` check is removed:
  Load() defaults empty to "memory" earlier, so the old guard
  never fired.

Test changes:

- New `internal/config/dsn_test.go` locks the predicate across
  memory/empty/file/bare-path/postgres/mysql inputs (absorbs
  the cases from the deleted `internal/runtime/runtime_test.go`).
- `internal/config/transport_auth_matrix_test.go` flips the
  "Wave C" TODO to an assertion: dev DSN without the ack flag
  must fail; dev DSN *with* the flag must succeed; postgres DSN
  always succeeds. Adds a negative case for empty-DSN default.
- Three existing streamable_http config tests
  (`TestLoadStreamableHTTPAllowsEmptyAPIKey`,
  `TestLoadStreamableHTTPWithoutStaticAPIKey`,
  `TestLoadStreamableHTTPRequiresOIDCIssuer`) now set
  `MCP_ALLOW_DEV_BACKEND=1` to keep exercising the dev path.
- `MCP_ALLOW_DEV_BACKEND` added to the matrix test's per-case
  env reset so the flag does not leak between cells.

Out of scope for this commit — tracked as a follow-up on the
same branch:

- Legacy HTTP prod-default flip (allow/warn/deny).
- Audit durability prod-default flip (best_effort/fail_closed).
- Consolidated ADR covering the three prod-fail-closed defaults.

Why: the repo's production-readiness claim is that shared-service
deployments are safe by default. Silent memory fallback on
streamable_http undermines that; no operator should be able to
misconfigure the multi-process transport and have the server
boot anyway. Failing at Load() puts the error in the startup
log where it's visible.

Verified: `make check` green on wave-h/streamable-fail-closed
(fmt + vet + full test suite). Drift check on the test: I
reverted `config.go`'s new guard locally and reran the matrix
test — the three new negative cases
(streamable_http + dev-DSN + no flag) passed as "ok" instead
of failing, proving they only pass with the guard in place. I
also spot-checked that the streamable harness in tests/harness/
bypasses Load() entirely (it calls controlplane.Open("memory")
directly), so E2E parity tests are unaffected.
